🔒️ Replace `==` with `secrets.compare_digest` in `verify_user`
The previous version of this line used `==` to check whether
the hashed submitted password was the same as the one stored
in the database. This *might* be vulnerable to timing attacks.

Since a reasonable hash function is used throughout, I don't
think this will be an issue assuming the hashes are distributed
well. It might be a good teaching point to bring timing attacks
up as something to guard against though.